+++ src/constants.ts
@@ -1,3 +1,4 @@
 export const FXML_GLOB = '**/*.fxml';
+export const CSS_GLOB = '**/*.css';
 export const JAVA_GLOB = '**/*.java';
 export const EXCLUDE_GLOB = '**/node_modules/**';